refactor: consolidate Activity type, use same type everywhere
