+++ src/components/Activities/types.ts
@@ -1,4 +1,2 @@
-export {
-  type Activity,
-  ActivityObjectType,
-} from '../../hooks/useSessionActivities'
+export { ActivityObjectType } from '../../types/generated/types-orchestration'
+export type { Activity } from '../../types'
