+++ src/components/Form/types.ts
@@ -1,0 +1,7 @@
+export type { Activity } from '../../types'
+
+export {
+  type AnswerInput,
+  FormDisplayMode,
+  type QuestionRuleResult,
+} from '../../types/generated/types-orchestration'
