+++ src/hooks/activityNavigation/types.ts
@@ -1,1 +1,1 @@
-export type { Activity } from '../../hooks/useSessionActivities'
+export type { Activity } from '../../types'
