+++ src/hooks/useChecklist/index.ts
@@ -1,3 +1,2 @@
 export { useChecklist } from './useChecklist'
-export { ActivityStatus } from './types'
 export type { Activity } from './types'
