+++ src/hooks/useChecklist/types.ts
@@ -1,10 +1,6 @@
-export {
-  ActivityStatus,
-  useGetChecklistQuery,
-  useSubmitChecklistMutation,
-} from '../../types/generated/types-orchestration'
+export { useGetChecklistQuery } from '../../types/generated/types-orchestration'
 
-export type { Activity } from '../../types/generated/types-orchestration'
+export type { Activity } from '../../types'
 export interface ChecklistItem {
   id: string
   label: string
