+++ src/hooks/useForm/types.ts
@@ -3,7 +3,6 @@ export type {
   Question,
   GetFormResponseQuery,
   GetFormQuery,
-  Activity,
 } from './../../types/generated/types-orchestration'
 
 export {
@@ -13,4 +12,5 @@ export {
   GetFormDocument,
 } from './../../types/generated/types-orchestration'
 
+export type { Activity } from '../../types'
 export type Form = GetFormQuery['form']['form']
