+++ src/hooks/useMessage/types.ts
@@ -2,13 +2,8 @@ import type { GetMessageQuery } from '../../types/generated/types-orchestration'
 
 export {
   useGetMessageQuery,
-  MessageFormat,
-  ActivityObjectType,
   useMarkMessageAsReadMutation,
 } from './../../types/generated/types-orchestration'
-export type {
-  AnswerInput,
-  Activity,
-} from './../../types/generated/types-orchestration'
 
+export type { Activity } from '../../types'
 export type Message = GetMessageQuery['message']['message']
