+++ src/hooks/useSubmitChecklist/index.ts
@@ -1,3 +1,2 @@
 export { useSubmitChecklist } from './useSubmitChecklist'
 export type { Activity } from './types'
-export { ActivityStatus } from './types'
