+++ src/hooks/useSubmitChecklist/types.ts
@@ -1,7 +1,3 @@
-export {
-  ActivityStatus,
-  useGetChecklistQuery,
-  useSubmitChecklistMutation,
-} from '../../types/generated/types-orchestration'
+export { useSubmitChecklistMutation } from '../../types/generated/types-orchestration'
 
-export type { Activity } from '../../hooks/useSessionActivities'
+export type { Activity } from '../../types'
