+++ src/hooks/useSubmitForm/types.ts
@@ -4,4 +4,4 @@ export {
 } from './../../types/generated/types-orchestration'
 export type { AnswerInput } from './../../types/generated/types-orchestration'
 
-export type { Activity } from '../../hooks/useSessionActivities'
+export type { Activity } from '../../types'
