+++ src/types/Activity.ts
@@ -1,0 +1,1 @@
+export type { Activity } from '../hooks/useSessionActivities'
