+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export type { Activity } from './Activity'
+export type { Maybe } from './Maybe'
