+++ types/index.ts
@@ -1,4 +1,3 @@
 export type { StartHostedActivitySessionParams } from './StartHostedActivitySessionParams'
 export type { StartHostedActivitySessionPayload } from './StartHostedActivitySessionPayload'
-export type { Maybe } from './Maybe'
 export { environment } from './Environment'
